# Fix #1930: Vhdl Identifier Length Cap