Handles/Fixes #1930 

Changes the scanning functions from fixed size strings to Unbounded_Strings as the LRM does allow for arbritrary identifier lengths rather than the current 1K/4K cap and does not meet standard that way.

Only the strings and their access have been changed, any code around it (such as keeping track of the Length of the string) has remained unmodified for now.

Since this change from fixed size strings to unbounded ones can have a bigger impact on performance, memory and so on I would highly appreciate taking a closer look before merging as this is my first contribution and Ada is not my strong suit.

The unlimited length works on my end, but it would be great if somebody could countercheck with a bigger project and see if everything works just fine.