@tgingold 2023-07-17
The patch is correct, but I have some doubts about the performances.
An unbounded string is heavy, and this is an issue in the scanner.  An identifier is a very common token and each time an identifier is scanned, malloc/free will be called.

I suppose we need to create an ad-hoc type, more efficient than Unbounded_Strings.
(There is also grt-vstrings that could be used).

Did you do any performance evaluation ? (Just asking).