@alpyen 2023-07-18
> The patch is correct, but I have some doubts about the performances. An unbounded string is heavy, and this is an issue in the scanner. An identifier is a very common token and each time an identifier is scanned, malloc/free will be called.

Yes, I was thinking about the same issue and wanted to see if this has any meaningful impact first.

> I suppose we need to create an ad-hoc type, more efficient than Unbounded_Strings. (There is also grt-vstrings that could be used).

Looking at the grt-vstrings it seems like you can allocate a a bigger chunk of memory instead of doing it bytewise to save on the mallocs/and frees, is that correct? If so, this would be a good substitution for unbounded strings. I didn't have it on my radar but I could try an implementation with that and see where it goes?

> Did you do any performance evaluation ? (Just asking).

Nothing of substance, the example file with a 1MB signal name and inverting it every few ns compiles and executes a few milliseconds measurably faster but it doesn't seem to scale with simulation runtime.
However I don't have any bigger project to compile against and see how big the performance hit actually is.