@umarcor 2023-07-18
I believe we will need some specific test which is particularly "heavy" with regard to the number of identifiers used in the design. According to CI, the impact of this PR is none.

- This PR took 36min 26s: https://github.com/ghdl/ghdl/actions/runs/5578998055
- Previous executions of the master branch took: 37m 49s, 45m 32s, 43m 27s, 37m 41s...

Since this PR was executed once only, I would consider the fact it is 1 min faster just incidental. CI has a variability of +-5min over a 40min average.
Anyway, it seems to not have a negative impact when running all of the tests in the testsuites. I believe that's meaningful considering the dialogue above.