@alpyen 2023-07-19
Agreed, I'm setting up the unisim compilation and will compare the three different versions (fixed size, unbounded, vstr) with one another once I have implemented the vstr one and will report back with benchmarks once that's done.