@alpyen 2023-07-21
Assuming my implementation is correct (last commit), the results are quite surprising.
I let my testfile (this time with a 5M signal name) and unisim (Xilinx 2022.2) compile (as default with mcode), these are the results:

```´
averaged over 5 runs for each mode using "time"
mode: real, user, system time

== Fixed Size Strings ==

example.vhdl:
-i: 152.0ms, 138.0ms, 13.6ms
-a: 152.2ms, 135.4ms, 17.0ms
-r: 695.5ms, 643.2ms, 52.0ms

unisim:
run 1: 6m30s260ms, 6m22s830ms, 10s330ms
run 2: 6m29s530ms, 6m22s870ms, 09s870ms


== Unbounded Strings ==

example.vhdl:
-i: 196.4ms, 175.4ms, 20.8ms
-a: 201.0ms, 173.8ms, 27.4ms
-r: 849.4ms, 758.0ms, 91.2ms

unisim:
run 1: 7m04s410ms, 6m56s970ms, 10s630ms
run 2: 7m05s560ms, 6m58s400ms, 10s300ms


== VStrings ==

example.vhdl:
-i: 328.4ms, 310.0ms, 18.4ms
-a: 327.6ms, 308.2ms, 19.2ms
-r: 874.6ms, 811.2ms, 63.2ms

unisim:
run 1: 7m04s520ms, 6m54s620ms, 13s080ms
run 2: 7m05s010ms, 6m54s700ms, 13s510ms
```
Unbounded_Strings seem to be faster on my machine than the Vstrings implementation.
The chunked allocation resulting in lower mallocs/frees shows up in the lower system time,
however it looks like the additional software overhead(?) of the Vstrings results in much higher user time.

Unisim times however are almost identical, the identifiers used there are probably much shorter
and probably therefore don't need to be extended further than the initial 32 byte vstring length.

Seeing the much higher time for Vstrings I decided to change the backend to LLVM and see if a different compiler will yield different results and the fixed size version was the only test to complete (after around 7 minutes).
Both Unbounded_Strings and Vstrings did not finish. The Unbounded one was still 'running' after 45 minutes before I exited out, and the Vstrings version flat out crashed in the beginning.