@tgingold 2023-08-16
(Back from being off).  I will investigate the llvm failure, this is weird.