@tgingold 2023-09-30
I have merged the part for trans, using vstrings, after adjustments.