@tgingold 2023-10-02
I have also modified and merge the change for the scanner. 

Closing this MR.